Step 6 - Introducing a read-only repository base interface.
Introduced ReadOnlyRepository to capture all read-only repository methods of CrudRepository and PagingAndSortingRepository. It also fixes the the type of the ID property to Long. Added Product domain class, sample data as well as an extension of ReadOnlyRepository to declare a query method to find Product instances containing a given description.